4.2 DB field type mod

+++ db/migrate/20160829235019_change_price_data_type_in_products.rb
@@ -1,5 +1,5 @@
 class ChangePriceDataTypeInProducts < ActiveRecord::Migration[5.0]
   def change
-  	change_column :products, :price, :decimal
+  	change_column :products, :price, :integer
   end
 end
